Refactor: store interactive element nodes in display.menu
Instead of creating these nodes right when showing a certain part of the menu, I decided to create them when initializing the menu module.
This ensures these nodes exist in the menu module without needing to be currently loaded in the html document. This way, other modules could access these nodes with getter methods when, for example, adding event listeners.