Update game state according to exact gameboard state
Before, the game state was updated assuming that a new move was made, so the gameState variables that determine the outcome could only be given a new different value other than its default values. That meant that if a move was undone, the gameState wouldn't update accordingly because there was no code to set the variables to their original values.